+++ odoo/songs/install/accounting.py
@@ -351,6 +351,15 @@ def set_esb_references(ctx):
 
 
 @anthem.log
+def import_account_payment_term(ctx):
+    """ Importing account payment term """
+    content = resource_stream(req, 'data/install/account.payment.term.csv')
+    load_csv_stream(ctx, 'account.payment.term', content, delimiter=',')
+    lines = resource_stream(req, 'data/install/account.payment.term.line.csv')
+    load_csv_stream(ctx, 'account.payment.term.line', lines, delimiter=',')
+
+
+@anthem.log
 def main(ctx):
     """ Configuring accounting """
     configure_missing_chart_of_account(ctx)
@@ -369,3 +378,4 @@ def main(ctx):
     settings(ctx)
     setup_sequences(ctx)
     set_esb_references(ctx)
+    import_account_payment_term(ctx)